--- functions/time/localtime.c
@@ -1,28 +1,0 @@
-/* localtime( const time_t * )
-
-   This file is part of the Public Domain C Library (PDCLib).
-   Permission is granted to use, modify, and / or redistribute at will.
-*/
-
-#include <time.h>
-
-#ifndef REGTEST
-
-struct tm * localtime( const time_t * timer )
-{
-    return NULL;
-}
-
-#endif
-
-#ifdef TEST
-
-#include "_PDCLIB_test.h"
-
-int main( void )
-{
-    TESTCASE( NO_TESTDRIVER );
-    return TEST_RESULTS;
-}
-
-#endif
